# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Knowledge Core API",
        "description": "Дата обновления: 29.10.2025",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/news": {
            "get": {
                "tags": [
                    "Новости"
                ],
                "summary": "Получить список всех новостей",
                "description": "Для любого пользователя возвращает список всех новостей.",
                "operationId": "af0419b50948e61ee929bfc55913e3d8",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Some name"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Some description"
                                            },
                                            "status": {
                                                "type": "integer",
                                                "example": 0
                                            },
                                            "date": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-11-01 01:00:00"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Новости"
                ],
                "summary": "Создание новой новости",
                "description": "Для администратора дает возможность создать новую новость.",
                "operationId": "d5ff9cc79f04122e45aba5d3b6cad419",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Some name"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Some description"
                                            },
                                            "status": {
                                                "type": "integer",
                                                "example": 0
                                            },
                                            "date": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-11-01 01:00:00"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Some name"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Some description"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 0
                                        },
                                        "date": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-11-01 01:00:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/news/{news}": {
            "get": {
                "tags": [
                    "Новости"
                ],
                "summary": "Получить единичную новость",
                "description": "Для любого пользователя возвращает новость по её идентификатору (id).",
                "operationId": "34218e7771dec97d0099881b1ceea645",
                "parameters": [
                    {
                        "name": "news",
                        "in": "path",
                        "description": "id новости",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Some name"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Some description"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 0
                                        },
                                        "date": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-11-01 01:00:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Новости"
                ],
                "summary": "Обновить новость",
                "description": "Для администратора дает возможность редактировать любую новость по её идентификатору (id).",
                "operationId": "fdf1f1b9fa30f2fcf864354dc2748b8c",
                "parameters": [
                    {
                        "name": "news",
                        "in": "path",
                        "description": "id новости",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Some name for edit"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Some description for edit"
                                            },
                                            "status": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "date": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-12-01 00:00:00"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Some new name"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Some new description"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "date": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-12-01 00:00:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Новости"
                ],
                "summary": "Удалить новость",
                "description": "Для администратора дает возможность удалить любую новость по её идентификатору (id).",
                "operationId": "c015f5a7d0cc412ad5b5c9e81e00eae5",
                "parameters": [
                    {
                        "name": "news",
                        "in": "path",
                        "description": "id новости",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}